+++ crates/nostr-sdk/src/lib.rs
@@ -38,10 +38,6 @@ pub use nostr_relay_pool::{
 #[doc(hidden)]
 pub use nostr_signer::{self as signer, NostrSigner, NostrSignerType};
 #[doc(hidden)]
-#[allow(deprecated)]
-#[cfg(feature = "sqlite")]
-pub use nostr_sqlite::SQLiteDatabase;
-#[doc(hidden)]
 #[cfg(feature = "nip57")]
 pub use nostr_zapper::{self as zapper, *};
 #[doc(hidden)]