+++ CHANGELOG.md
@@ -100,6 +100,7 @@
 * pool: remove `RelayPool::reconcile_advanced` ([Yuki Kishimoto])
 * pool: remove `RelayPool::reconcile_with_items` ([Yuki Kishimoto])
 * webln: remove `nostr-webln` crate ([Yuki Kishimoto])
+* sqlite: remove `nostr-sqlite` crate ([Yuki Kishimoto])
 
 ## [v0.35.0]